+++ tests/harness.lua
@@ -162,6 +162,7 @@ load("Spec_Arms.lua")
 load("Spec_Outlaw.lua")
 load("Spec_Elemental.lua")
 load("Spec_Devourer.lua")
+load("Spec_BeastMastery.lua")
 
 PRIO.Engine:OnSpecChanged()
 H.spec = PRIO.specs[269]
@@ -169,6 +170,7 @@ H.armsSpec = PRIO.specs[71]
 H.outlawSpec = PRIO.specs[260]
 H.eleSpec = PRIO.specs[262]
 H.devourerSpec = PRIO.specs[1480]
+H.bmSpec = PRIO.specs[253]
 H.Engine = PRIO.Engine
 H.Cond = PRIO.Cond
 H.API = API